refactor(dsp): rename non-DSP module files by actual function
- afe/rnnoise/core/dsp.ts -> constants.ts: contains RNNOISE_FRAME_SIZE,
  vadToProbability + dsp re-exports (no DSP implementation).
- few-shot/core/dsp.ts -> quality.ts: cosine similarity + sample-quality
  checks (quality gating, not DSP).
- Update all importers (engine.ts, index.ts, playground.tsx, FewShotEngine,
  tests) + spec.json l1 paths + vitest config comment + module-spec.md
  example + ADR-032 note.
- afe/graph/core/dsp.ts (real viz spectrum) and kws/engine/core/dsp.ts
  (decision logic) keep their names.
- Full-repo typecheck/lint/unit green.
