+++ packages/modules/afe/rnnoise/vitest.config.ts
@@ -2,7 +2,7 @@ import { defineConfig } from 'vitest/config'
 
 /**
  * RNNoise module vitest config (ADR-026).
- *   - L1: pure DSP (tests/dsp.test.ts)
+ *   - L1: domain constants + VAD mapping (tests/constants.test.ts)
  *   - L2: wasm runtime in Node (tests/wasm-runtime.test.ts) - the vendored
  *     emscripten glue supports ENVIRONMENT_IS_NODE, so the same artifact that
  *     boots in the browser boots here, in seconds.
